Fix: make jobtype_root_package_data_roots a tuple
This fixes a bug where get_package_data may get into what seems like an
endless loop to the user.  When jobtype_root_package_data_roots is a
string, get_package_data() will eventually be called with the string
"pyfarm/" as the argument.  Instead of looping over several directories,
it will now loop over the string "pyfarm/".  Since the last character in
that is "/", it will eventually call os.walk() with "/", causing it to
try to include the entire VFS into the package_data, including stuff
like /dev, /sys and /proc.