Updating stdout/stderr handling, added docs, improved code flow
This change updates our existing methods for handling output from a process to:

* Handle both stderr and stdout
* Call a single private method from the job type so the job type contains the
  top level logic
* Use this new single private method to send the correct data and calls to a
  public method which will then handle the rest of the
* Provide indiviaul formatting for both stdout and stderr
* Provide a single entrypoint for individual line processing so it's easiear for
  an implementor to modify and understand

There may be a few small deviations from the original discussion here but it's
more than likely unintentional.  This change is mainly meant to improve on
the code flow a bit, add documentation, and keep from repeating ourselvles where
possible.  The idea is to offer the same or better flexibility than before while
being more complete and easiear to understand.