Allow JobType.stop() to be called several times
stop() just send a signal to all child processes.  There is no hard
reason for not calling it any number of times.  This is particularly
important if it is called at a moment when a chil process has already
been scheduled to start but isn't running yet, in which case the call
would be ineffective.