Call treq.put with persistent=False
This is a workaround to make some of the unit tests work again. All
network connections need to be closed when a unit test ends, but under
normal conditions, treq will use keep-alive on HTTP-Connections, keeping
them open even after an HTTP request.

A solution that affects only the tests themselves and not the normal way
the code works would be preferred to this one.