Use factory if provided in BaseHTTPTestCase
This allows test cases inheriting from this to specify a factory for
instancing the resource to be tested.  This is useful for cases where
just calling <class>() with no arguments is not going to work, because
the class's constructor needs an argument.  The Assign resource is an
example for that.