Don't use inlineCallbacks on schedule_call due to recursion issues.
In the case of inlineCallbacks there's a recursion problem with
schedule_call calling itself over and over.  It does not
actually cause a recursion problem while running however on
exit Twisted has a problem unwinding the full stack of deferreds.

This code would run fun until Ctrl+C is pressed resulting
in a problem inside of Twisted that looks like a recursion
issue:

  from twisted.internet import reactor
  from pyfarm.agent.service import Agent

  agent = Agent()

  def shutdown():
    agent.shutting_down = True

  agent.repeating_call(0, lambda: None)
  reactor.callLater(3, shutdown)
  reactor.run()

In the case of schedule_call we don't really care
about the result so inlineCallbacks in this case is
not 100% necessary (though it would have been nice
to keep).