Use pseudostream "internal" for internal messages
When logging internal messages to the task log, the agent would log it
under one of the stream STDOUT or STDERR. This is arguably wrong,
though, since that indicates that those lines have been read from the
stdout or stderr, respectively, of th child process.

This changes that to use the special string "internal" in place of a
stream number, so that this message cannot be confused as coming from
child process.