Fix: Reannounce before removing stopped assignment
The previous commit would make the agent update its state with the
master after removing the the finished assignment from
current_assignments, but potentially before update the states of the
individual tasks to done or failed.  This would make the master assume
the assignment was failed, because the tasks were not present in the
agent's list of tasks during reannounce.