Removing the top level loader, adding EncodedStringTemplate().
This commit removes the top level Loader() class.  This class
was not really anything more than an import guard woud provide
us.  This commit also adds in the EncodedStringTemplate class
which will ensure that we always return a string from .render().
Twisted requires this because it can't seem to handle unicode output
which is what jinja2.Template.render will return by default.