bugfix, mixup with self.started_deferred/self.start_deferred
Currently the Process object sets and uses self.started_deferred. We
however have a self.start_deferred, like we have a self.stopped_deferred,
property that is not in use.  This change renames the existing property
to the proper name so both self.started_deferred and self.stopped_deferred
benefit from the same checks.  Tests are being written to prove this has
not broken anything from a functional standpoint.