don't encode to ASCII, DummyRequest.write seems to get it wrong
This rolls back a previous commit which was encoding output
from utility.dumps to ascii.  That change was made because
DummyRequest.write raised a TypeError if the input variable
was anything but a bytes object.  That behavior however is
not replicated in the Request.write method and unicode would
instead be cast to 'bytes'.