change default header values for APIResource
In the past there have been cases where we allowed
headers like Content-Type and Accept to be unset for
API requests.  This change readds this support (which
also unbreaks a few tests).  The reasoning behind this
is two parts:
  * An incoming request to the API we generally should
    assume is json.  Because of this, we'll just try
    to decode the request anyway.
  * An outgoing reponse is assumed to be json

In any case, if somone passes a value for Accept or
Content-Type that we don't explicitly allow we will
still reject the request.