Rework how LoggerPool.flush() writes to disk
Simplifies LoggerPool.flush() a bit.  Now only one thread will ever
write messages to the log file at a time, until it is finished.
Before, if there were several threads writing (which can happen if a lot
of output is produced in a short time), the lock would constantly pass
from one thread to the next, and as long as messages keep coming in, no
thread will ever actually finish.

Also, now the log file is unconditionally flushed every time a thread
ends.  This is better for jobtypes that produce little output in case an
admin wants to find out how far a running assignment has gotten (or why
it is stuck).