change changed() logging to be debug
The agent's configuration logger is probably
one of the most verbose loggers.  Its output
is important but probably only in cases where
there's a problem being debugged.  Even then the
configuration is usually not going to be the cause
of the issue so keeping the configuration logger
in debug only mode probably makes more sense.