Configurably reject multiple simultaneous batches
This makes agent sharing configurable. By default, all new assignments
will be rejected if we still have at least one assignment with at least
one task that is not done or failed.

This also slightly changes the semantics of the /assign endpoint:
The agent will now return 202 ACCEPTED if it is sent an assignment that
is identical to one we already have, but won't internally register and
start a new jobtype for it. There is now no longer a way for a client
to tell the difference between that and a fresh assignment, effectively
making the endpoint idempotent.

This is necessary because 409 CONFLICT is now a hard error that means
the assignment will NOT be executed and that the master will have to
react to. (Usually by finding a new node for the rejected task ids.)