Reject assignments if we are busy reannouncing
Accepting assignments during a running reannounce can lead to weird
race, where the reannounce request is locally started right before the
incoming assignment, and uses data from that point in time, but doesn't
actually reach the master until after the incoming assignment. At that
point, the master has already marked the tasks as sent to agent and will
assume we failed it because our announce is lacking those task ids.