Do not treat insufficient temp space as failure
When a task fails because we couldn't free enough disk space for it,
that doesn't mean there is anything inherently wrong with that task, it
only mean that task cannot be run on this node at this point in time.

If we only set the tasks state to queued and dissociate it with this
agent, however, there is a high likelihood that the master's scheduler
will just immediately reassign that task to us.  To avoid that, we add
this agent to the failed_on_agents list of that task.