creating the APITestServerResource context manager
The class APITestServerResource is a context manager
which can setup a test server on a random port.  This
allows a test to very simply use this class for building
simplified tests where an HTTP response is required:

    @inlineCallbacks
    def test_download(self):
        classname = "AgentUnittest" + os.urandom(8).encode("hex")
        cache = Cache()

        with APITestServer(
            "/jobtypes/%s/versions/1" % classname,
            headers={"Content-Type": "application/json"},
            code=OK, response=json.dumps({"data": "This is a job type"})
        ):
            response = yield cache._download_jobtype(classname, 1)

Eventually this context manager will replace the setup in other tests
since it's more generalized and requires less work to setup and use.