refactor to use/cache job type dictionary rather than just the source (tests not updated yet)
Because we need the job type's class name, and the master does
not provide this, we need to cache and use the job type dictionary
directly.  I was hoping to reduce bandwidth and just use the code
directly but this will be unreliable.  Going this route also has
the added benefit of caching the full job type information including
the software requirements at the time.  We're not using this information
right now anyway but it could be useful in the future.