raise TypeError if jobtype_default_environment is not a dictionary
This is a potentially breaking change.  Before this we would
simply log a warning if the configuration defined jobtype_default_environment
and it was not a dictionary.  From the perspective of job type
implementor this would have been a silent failure and resulted in
potentially unexpected behavior within a running job.